Move validate_payload into SystemDatabaseReader
